+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,18 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* `az aks nodepool update`: clean up some useless code in the update managed gpu function. 
+* `az aks machine add`: Add `--spot-max-price` flag support to set the max price (in US Dollars) you are willing to pay for spot instances on a machine.
+* `az aks machine add`: Add `--eviction-policy` flag support to set the eviction policy for a machine.
+* `az aks machine add`: Add `--enable-ultra-ssd` flag support to enable ultra ssd on a machine.
+
+20.0.0b1
++++++++
+* [Breaking Change] `az aks create/update`: Change `--nat-gateway-outbound-ips` and `--nat-gateway-outbound-ip-prefixes` to use comma-separated values, consistent with load balancer outbound IP parameters.
+
+19.0.0b30
++++++++
+* Add option `AzureContainerLinux` to `--os-sku` for `az aks create`, `az aks nodepool add`, and `az aks nodepool update`.
 
 19.0.0b29
 +++++++